src/auth/utils.py
from passlib.context import CryptContext
from datetime import datetime, timedelta
from src.config import Config
import jwt
import uuid
import logging


passwd_context = CryptContext(schemes=['bcrypt'])
EXPIRE_ACCESS_TOKEN = 3600

def generate_passwd_hash(passwd:str) ->str:
    hash = passwd_context.hash(passwd)

    return hash

def verify_password(passwd:str, hash:str) ->bool:
    return passwd_context.verify(passwd, hash)

def create_access_token(user_data:dict, expiry:timedelta=None, refresh:bool=False):
    payload = {}
    expire = datetime.now() + (expiry if expiry is not None else timedelta(seconds=EXPIRE_ACCESS_TOKEN))

    payload['user'] = user_data
    payload['exp'] = expire
    payload['jti'] = str(uuid.uuid4())

    payload['refresh'] = refresh

    token = jwt.encode(payload=payload, key=Config.JWT_SECRET, algorithm=Config.JWT_ALGORITHM)

    return token

def decode_token(token:str) ->dict:
    try:
        token_data = jwt.decode(jwt=token, key=Config.JWT_SECRET, algorithms=[Config.JWT_ALGORITHM])

        return token_data
    except jwt.PyJWTError as e:
        logging.exception(e)

